# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: University Faculty Work Portal API
  description: |
    API для системи планування, звітування та оцінювання роботи викладачів університету.
    
    **Типи користувачів:**
    - GUEST - незареєстрований користувач (публічний доступ)
    - TEACHER - викладач
    - DIRECTOR - завідувач кафедри/декан (має права викладача + додаткові)
    - ADMIN - адміністратор системи
    
    **Автентифікація:**
    - JWT tokens (Access: 30 хв, Refresh: 30 днів)
    - Hybrid strategy з Redis кешуванням
  version: 1.0.0
  contact:
    name: API Support
    email: api@stu.cn.ua
servers:
  - url: https://npp.stu.cn.ua/api/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        uuid:
          type: string
          format: uuid
        firstName:
          type: string
          example: "Олександр"
        lastName:
          type: string
          example: "Шевченко"
        patronymic:
          type: string
          example: "Миколайович"
        login:
          type: string
          example: "shevchenko.o.m"
        email:
          type: string
          format: email
          example: "shevchenko@stu.cn.ua"
        role:
          type: string
          enum: [TEACHER, DIRECTOR, ADMIN]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        details:
          $ref: '#/components/schemas/UserDetailsResponse'
        departments:
          type: array
          items:
            $ref: '#/components/schemas/UserDepartmentResponse'
...
    UserDetailsResponse:
      type: object
      properties:
        id:
          type: integer
        photo:
          type: string
          nullable: true
          description: URL фотографії
        degree:
          $ref: '#/components/schemas/DictionaryItemResponse'
        position:
          $ref: '#/components/schemas/DictionaryItemResponse'
        englishLevel:
          $ref: '#/components/schemas/DictionaryItemResponse'
        nasAffiliation:
          $ref: '#/components/schemas/DictionaryItemResponse'
        honors:
          $ref: '#/components/schemas/DictionaryItemResponse'
    UserDepartmentResponse:
      type: object
      properties:
        department:
          $ref: '#/components/schemas/DepartmentResponse'
        joint:
          type: integer
          nullable: true
          description: "Суміщення ставок"
        workload:
          type: number
          format: float
          nullable: true
          description: "Навантаження"
...
    CreateUserRequest:
      type: object
      required:
        - firstName
        - lastName
        - patronymic
        - email
        - role
      properties:
        firstName:
          type: string
          example: "Олександр"
        lastName:
          type: string
          example: "Шевченко"
        patronymic:
          type: string
          example: "Миколайович"
        email:
          type: string
          format: email
          example: "shevchenko@stu.cn.ua"
        role:
          type: string
          enum: [TEACHER, DIRECTOR, ADMIN]
...
    UpdateUserDetailsRequest:
      type: object
      properties:
        degreeId:
          type: integer
        positionId:
          type: integer
        englishLevelId:
          type: integer
        nasAffiliationId:
          type: integer
        honorsId:
          type: integer
    AssignDepartmentRequest:
      type: object
      required:
        - departmentId
      properties:
        departmentId:
          type: integer
        joint:
          type: integer
          nullable: true
        workload:
          type: number
          format: float
          nullable: true
    UpdateDepartmentAssignmentRequest:
      type: object
      properties:
        joint:
          type: integer
          nullable: true
        workload:
          type: number
          format: float
          nullable: true

    # ==================== ACTIVITY SCHEMAS ====================
...
    DepartmentResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        depLogo:
          type: string
        depHead:
          type: string
        depPage:
          type: string
        active:
          type: boolean
        faculty:
          $ref: '#/components/schemas/FacultyResponse'
...
    FacultyResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        facLogo:
          type: string
        facHead:
          type: string
        facPage:
          type: string
        active:
          type: boolean
        institute:
          $ref: '#/components/schemas/InstituteResponse'
...
    InstituteResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        instLogo:
          type: string
        instHead:
          type: string
        instPage:
          type: string
        active:
          type: boolean
...
    DictionaryItemResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
          description: "Тільки для degrees"
        level:
          type: string
          description: "Тільки для english-levels"
        active:
          type: boolean
...
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
paths:
...
  /users/{id}/departments:
    get:
      tags:
        - User Profile
      summary: Отримання кафедр користувача
      description: |
        Отримання списку кафедр користувача з інформацією про суміщення та навантаження.
        **Доступ:**
        - Власний профіль (бачить свої дані)
        - ADMIN (бачить будь-якого користувача)
        - DIRECTOR (бачить викладачів своєї кафедри)
      operationId: getUserDepartments
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список кафедр користувача
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDepartmentResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Користувача не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - User Profile
      summary: Призначення на кафедру
      description: |
        Додавання викладача на кафедру з вказанням суміщення та навантаження.
        **Доступ:** тільки ADMIN
      operationId: assignDepartment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignDepartmentRequest'
      responses:
        '201':
          description: Викладача призначено на кафедру
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Користувача або кафедру не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/departments/{departmentId}:
    put:
      tags:
        - User Profile
      summary: Оновлення даних про кафедру
      description: |
        Оновлення суміщення та навантаження на кафедрі.
        **Доступ:** тільки ADMIN
      operationId: updateUserDepartment
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentAssignmentRequest'
      responses:
        '200':
          description: Дані оновлено
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Користувача, кафедру або призначення не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - User Profile
      summary: Видалення з кафедри
      description: |
        Видалення викладача з кафедри.
        **Доступ:** тільки ADMIN
      operationId: removeDepartment
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Викладача видалено з кафедри
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Користувача, кафедру або призначення не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== ACTIVITIES ENDPOINTS ====================
...
  /users:
    get:
      tags:
        - Users
      summary: Список користувачів
      description: |
        Отримання списку всіх користувачів.
        **Доступ:** тільки ADMIN
      operationId: getUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
            enum: [TEACHER, DIRECTOR, ADMIN]
        - name: active
          in: query
          schema:
            type: boolean
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список користувачів
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Users
      summary: Створення користувача
      description: |
        Створення нового користувача. Генерується токен, відправляється email.
        **Доступ:** тільки ADMIN
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Користувача створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Конфлікт - користувач з таким email або login вже існує
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/{id}/details:
    get:
      tags:
        - User Profile
      summary: Отримання деталей профілю
      description: |
        Отримання додаткових даних викладача (ступінь, посада, англійська, НАН, звання).
        **Доступ:** ADMIN або власний профіль
      operationId: getUserDetails
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Деталі профілю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailsResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Користувача не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - User Profile
      summary: Оновлення деталей профілю
      description: |
        Оновлення додаткових даних викладача (ступінь, посада, англійська, НАН, звання).
        **Доступ:** ADMIN або власний профіль (TEACHER/DIRECTOR можуть редагувати свої дані)
      operationId: updateUserDetails
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDetailsRequest'
      responses:
        '200':
          description: Деталі оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailsResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Користувача не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/photo:
    post:
      tags:
        - User Profile
      summary: Завантаження фото викладача
      description: |
        Завантаження фото профілю (multipart/form-data).
        **Доступ:** тільки власний профіль (TEACHER/DIRECTOR самі завантажують своє фото)
      operationId: uploadUserPhoto
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
              properties:
                photo:
                  type: string
                  format: binary
                  description: "Файл зображення (JPEG, PNG). Макс. 5MB"
      responses:
        '200':
          description: Фото успішно завантажено
          content:
            application/json:
              schema:
                type: object
                properties:
                  photoUrl:
                    type: string
                    description: URL завантаженого фото
                    example: "https://cdn.stu.cn.ua/photos/user_123.jpg"
        '400':
          description: Невалідний файл (формат або розмір)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
            description: Недостатньо прав (тільки власний профіль)
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '413':
            description: Файл занадто великий (більше 5MB)
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'